redo all layout
