+++ docs/src/wordValidator.js
@@ -283,3 +283,4 @@ export class WordValidator {
     return uniqueWordsAndPositions;
   }
 }
+
