+++ docs/styles/game-header.css
@@ -1,0 +1,37 @@
+
+.game-title {
+  /* Styles for the game title */
+  font-size: 1.8rem;
+  margin-bottom: 8px;
+}
+
+.score-container {
+  /* Styles for the score and letters container */
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+  gap: 8px;
+  margin-bottom: 8px;
+}
+
+.score-container h2 {
+  /* Styles for score and letters text */
+  font-size: 1rem;
+  margin: 0;
+}
+
+.start-btn {
+  /* Styles for the start button */
+  padding: 8px 16px;
+  font-size: 1rem;
+  background: #28a745;
+  color: white;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+  transition: background 0.2s;
+}
+
+.start-btn:hover {
+  background: #218838;
+}
