+++ docs/styles/styles.css
@@ -28,26 +28,23 @@
   }
 }
 
-
 /* Game Wrapper - Contains both header and game board */
-.game-wrapper {
-  max-width: 420px;
-  width: 100%;
-  height: 100vh;
+.all-wrapper {
+  max-width: min(80vw, 600px);
+  max-height: 80vh;
+  margin: 0 auto;
   display: flex;
   flex-direction: column;
   align-items: center;
-  margin: 0 auto;
 
 }
 
 /* Game Header - Takes 1/4 of the height */
 .game-header {
-  height: 25vh;
-  min-height: 80px;
-  max-height: 150px;
+  flex: 1;
+  min-height: 0;
   width: 100%;
-  max-width: inherit;
+  box-sizing: border-box;
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -56,7 +53,6 @@
   box-shadow: 0 2px 6px rgba(0,0,0,0.06);
   border-radius: 8px;
   margin-bottom: 8px;
-  box-sizing: border-box;
   text-align: center;
 }
 
@@ -95,19 +91,35 @@
 
 /* Main container - Contains game board and right panel - Takes 75% of the height */
 .main-container {
-  height: 75vh;
+  flex: 5;
+  min-height: 0;
   width: 100%;
   display: flex;
+  flex-direction: row;
   justify-content: center;
+  gap: 8px;
 }
+
+.game-board-wrapper {
+  flex: 2;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: flex-start;
+  gap: 8px;
+  width: 100%;
+  box-sizing: border-box;
+}
+
 /* Game grid - responsive grid */
 #game-grid {
+  flex: 6;
+  min-height: 0;
+  height: auto;
   display: grid;
   grid-template-columns: repeat(7, 1fr);
   grid-template-rows: repeat(6, 1fr);
   width: 100%;
-  max-width: 420px;
-  min-width: 0;
   aspect-ratio: 7 / 6; /* Force proper proportions: 7 cols by 6 rows */
   gap: 0;
   margin: 0 auto;
@@ -133,6 +145,9 @@
 .tile.board {
   background: #f8f8fa;
   box-shadow: 0 0 0 1px #444 inset;
+  height: 100%;
+  width: 100%;
+  aspect-ratio: 1/1;
 }
 
 /* Falling tile animation style */
@@ -149,10 +164,10 @@
 
 /* Tile Spawn Row styles */
 #tile-spawn-row {
+  flex: 1;
   display: grid;
   grid-template-columns: repeat(7, 1fr);
   width: 100%;
-  max-width: 420px;
   height: auto;
   aspect-ratio: 7 / 1;
   margin: 0 auto;
@@ -227,6 +242,9 @@
 
 /* Preview container styles */
 .preview-panel {
+  flex: 1;
+  min-height: 0;
+  height: auto;
   display: flex;
   width: 100%;
   margin: 0 auto;
@@ -246,10 +264,7 @@
   display: grid;
   flex: 5;
   grid-template-columns: repeat(5, 1fr); /* or 5, if you want 5 tiles */
-  width: 100%;           /* Fill parent width */
-  max-width: inherit;    /* Inherit from parent (.game-wrapper or .board-wrapper) */
-  gap: 1px;
-  margin: 8px 0;
+  height: 100%;           /* Fill parent width */
 }
 
 
@@ -299,14 +314,54 @@
 
 /* Inner container that holds preview, tile-spawn-row and the grid */
 .game-board-container {
+  display: grid;
+  grid-template-rows: 1fr 1fr 6fr;
+  grid-template-columns: repeat(7, 1fr);
+  aspect-ratio: 7 / 8;
+  width: 100%;
+  max-width: min(80vw, 600px);
+  margin: 0 auto;
+  box-sizing: border-box;
+  height: auto;
+}
+
+.preview-panel {
+  grid-row: 1 / 2;
+  grid-column: 1 / 8;
+}
+
+#spawn-row {
+  grid-row: 2 / 3;
+  grid-column: 1 / 8;
+}
+
+#game-grid {
+  grid-row: 3 / 9;
+  grid-column: 1 / 8;
+  display: grid;
+  grid-template-rows: repeat(6, 1fr);
+  grid-template-columns: repeat(7, 1fr);
+  width: 100%;
   height: 100%;
+}
+
+.tile.board {
+  aspect-ratio: 1 / 1;
+  width: 100%;
+  height: 100%;
+}
+
+.made-words-list {
+  flex: 1;
+  width: 100%;
+  box-sizing: border-box;
+  background: #e3f2fd;
+  border: 2px solid #90caf9;
+  border-radius: 8px;
+  padding: 16px;
+  margin-top: 0;
   display: flex;
   flex-direction: column;
   align-items: center;
-  justify-content: flex-start;
-  gap: 8px;
-  width: 100%;
-  box-sizing: border-box;
-  padding: 4px 0;
+  overflow-y: auto;
 }
-
