+++ docs/styles/tiles.css
@@ -1,0 +1,76 @@
+
+.tile.board {
+	box-sizing: border-box;
+	width: 100%;
+	aspect-ratio: 1 / 1;
+	outline: 1px solid #333;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+}
+
+/* Spawn row inactive: invisible border */
+.tile.spawn.inactive, .tile.preview.inactive {
+  border: 1px solid transparent;
+  opacity: 0;
+}
+
+/* Spawn row active: bold border */
+.tile.spawn.active, .tile.preview.active {
+  border: 1px solid #222;
+  opacity: 1;
+}
+
+.tile.spawn:hover {
+  background: rgba(173, 216, 230, 0.2); /* Light blue hint on hover */
+  cursor: pointer;
+}
+
+/* Word completion animations */
+.tile.highlight {
+  background: #28a745 !important; /* Bootstrap success green */
+  color: white !important;
+  transform: scale(1.1);
+  box-shadow: 0 0 20px rgba(40, 167, 69, 0.6);
+  z-index: 10;
+  position: relative;
+}
+
+/* Shake animation for found words */
+.tile.shake {
+  animation: shake 0.6s ease-in-out;
+}
+
+@keyframes shake {
+  0%, 100% { transform: translateX(0) scale(1.1); }
+  10%, 30%, 50%, 70%, 90% { transform: translateX(-5px) scale(1.1); }
+  20%, 40%, 60%, 80% { transform: translateX(5px) scale(1.1); }
+}
+
+/* Additional pulsing effect for highlighted tiles */
+.tile.highlight.shake {
+  animation: shake-pulse 0.6s ease-in-out;
+}
+
+@keyframes shake-pulse {
+  0% { 
+    transform: translateX(0) scale(1.1);
+    box-shadow: 0 0 20px rgba(40, 167, 69, 0.6);
+  }
+  25% { 
+    transform: translateX(-5px) scale(1.15);
+    box-shadow: 0 0 25px rgba(40, 167, 69, 0.8);
+  }
+  50% { 
+    transform: translateX(5px) scale(1.1);
+    box-shadow: 0 0 30px rgba(40, 167, 69, 1.0);
+  }
+  75% { 
+    transform: translateX(-3px) scale(1.15);
+    box-shadow: 0 0 25px rgba(40, 167, 69, 0.8);
+  }
+  100% { 
+    transform: translateX(0) scale(1.1);
+    box-shadow: 0 0 20px rgba(40, 167, 69, 0.6);
+  }
+}
